Move EmailPreview into a folder
